% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Monte_Carlo_Inputs.R
\name{Generate_Monte_Carlo_Inputs}
\alias{Generate_Monte_Carlo_Inputs}
\title{Generate Monte Carlo Inputs}
\usage{
Generate_Monte_Carlo_Inputs(
  base_values,
  deviations,
  n,
  distribution = "normal"
)
}
\arguments{
\item{base_values}{dataframe which holds Monte Carlo base values}

\item{deviations}{dataframe containing deviations analysis}

\item{n}{number of iterations for analysis}

\item{distribution}{either "normal or "uniform"}
}
\value{
a list of dataframes with the purturbed values with the same
 structure as base_value
}
\description{
Generates inputs for a Monte Carlo analysis.
}
\details{
The inputs for a Monte Carlo Analysis are perturbed around a base value. 
 
This command accepts a base_value and a deviations dataframe then returns
a list of dataframes with elements perturbed according to deviations.

For a normal distribution the deviation is the standard deviation. 
For a uniform distribution the deviation is the absolute deviation.
}
\examples{
base_values = list(
   speed = c(10, 20),
   position = c(30, 40)
)

deviations = list(
  speed = c(0.5, 50),
  position = 2
)

Generate_Monte_Carlo_Inputs(base_values, deviations, 2)

}
